// tmp-harec-verify/src/lib/types.ts
export interface Question {
  id: string;
  examPartId: string;
  sectionId: string;
  wording: string;
  choices: Choice[];
  correctChoiceId: string;
  explanation: string;
  relatedLessonIds: string[];
  tags: string[];
}

export interface Choice {
  id: string;
  text: string;
}
// ...
export interface QuestionStat {
  questionId: string;
  timesAnswered: number;
  timesCorrect: number;
  lastAnswered: string;
  markedDifficult: boolean;
  consecutiveCorrect: number;
  nextReviewDate?: string;
}